Use a temp work dir when installing
Create and automatically clean-up a temporary directory.

This makes the process compliant with asdf requirements  to avoid
writing to `ASDF_INSTALL_PATH` until the download, extraction and build
are done to avoid racing with other processes trying to access the JDK.
https://github.com/asdf-vm/asdf/blob/master/docs/creating-plugins.md#bininstall

Since the directory is automatically cleaned up on exit, functions that
extract the archives don't need to worry about cleaning up individual
files as long as they are written to the working directory.

Also fix a problem with word splitting of `versionstr` - if unquoted,
the value is split by the shell before passing it to `read` which meant
that `distro` was actually being assigned DISTRO SPACE VERSION and
version was an empty string, it only worked further down coincidentally
because the variables weren't quoted so `distro` was subject to word
splitting before being passed to `get_java` and `unpack_java`.